requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pymongo>=4.5.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.1